# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Packaging Compliance AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      tags: [storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for direct file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /regulations:
    get:
      operationId: listRegulations
      tags: [regulations]
      summary: Browse the regulatory knowledge base
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: agency
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of regulations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Regulation"
    post:
      operationId: createRegulation
      tags: [regulations]
      summary: Add a regulation to the knowledge base
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegulationInput"
      responses:
        "201":
          description: Created regulation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Regulation"
  /regulations/{id}:
    get:
      operationId: getRegulation
      tags: [regulations]
      summary: Get a single regulation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Regulation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Regulation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers and vendors
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Add a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get a supplier with its packages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Package:
      type: object
      properties:
        id:
          type: integer
        sku:
          type: string
        upc:
          type: ["string", "null"]
        name:
          type: string
        brand:
          type: string
        vendor:
          type: string
        category:
          type: string
        country:
          type: ["string", "null"]
        netWeight:
          type: ["string", "null"]
        dimensions:
          type: ["string", "null"]
        packageType:
          type: ["string", "null"]
        productType:
          type: ["string", "null"]
        manufacturingRegion:
          type: ["string", "null"]
        status:
          type: string
        grade:
          type: ["string", "null"]
        riskScore:
          type: ["integer", "null"]
        complianceStatus:
          type: string
        reviewer:
          type: ["string", "null"]
        artworkUrl:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        criticalCount:
          type: integer
        majorCount:
          type: integer
        minorCount:
          type: integer
        analyzedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - sku
        - name
        - brand
        - vendor
        - category
        - status
        - complianceStatus
        - criticalCount
        - majorCount
        - minorCount
        - createdAt
        - updatedAt
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
...
    Regulation:
      type: object
      properties:
        id:
          type: integer
        agency:
          type: string
        category:
          type: string
        ruleCode:
          type: string
        title:
          type: string
        summary:
          type: string
        regulationText:
          type: ["string", "null"]
        section:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        publicationDate:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - agency
        - category
        - ruleCode
        - title
        - summary
        - createdAt
    RegulationInput:
      type: object
      properties:
        agency:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        ruleCode:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        summary:
          type: string
          minLength: 1
        regulationText:
          type: string
        section:
          type: string
        source:
          type: string
        publicationDate:
          type: string
      required:
        - agency
        - category
        - ruleCode
        - title
        - summary
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        riskLevel:
          type: string
        contactEmail:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        complianceScore:
          type: integer
        packagesReviewed:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - riskLevel
        - complianceScore
        - packagesReviewed
        - createdAt
    SupplierDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        riskLevel:
          type: string
        contactEmail:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        complianceScore:
          type: integer
        packagesReviewed:
          type: integer
        createdAt:
          type: string
          format: date-time
        packages:
          type: array
          items:
            $ref: "#/components/schemas/Package"
      required:
        - id
        - name
        - riskLevel
        - complianceScore
        - packagesReviewed
        - createdAt
        - packages
    SupplierInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        code:
          type: string
        category:
          type: string
        riskLevel:
          type: string
        contactEmail:
          type: string
        country:
          type: string
      required:
        - name
...